+++ .changeset/two-rabbits-move.md
@@ -1,0 +1,5 @@
+---
+'rabbit-mq11111': minor
+---
+
+Two functions -> micro/saga -> it creates the consumers and returns an emitter
