+++ packages/RabbitMQ/src/Consumer/channels/index.ts
@@ -1,0 +1,2 @@
+export * from './Consume';
+export * from './Microservice';
