+++ packages/RabbitMQ/src/Consumer/index.ts
@@ -1,3 +1,5 @@
-export * from './consumeWithEmitter';
+export * from './callbacks';
+export * from './channels';
+export * from './consume';
 export * from './create';
 export * from './nack';
