+++ packages/RabbitMQ/src/constants.ts
@@ -1,3 +1,5 @@
+export const REPLY_TO_SAGA_QUEUE = 'reply_to_saga';
+
 export const REQUEUE_EXCHANGE = 'requeue_exchange';
 export const NACKING_DELAY_MS = 2000;
 // some comment here to test the behavior of the snippet
