+++ packages/RabbitMQ/src/index.ts
@@ -1,4 +1,5 @@
 export * from './Broker';
 export * from './Connections';
+export * from './constants';
 export * from './Consumer';
 export * from './@types';
